moved ProviderSearchEventStatus from contract to model
